+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 * SSL_* (mod_ssl) CGI meta-variables are only set when the server is configured to set them.
 * The list of suffixes for resources with content type text/gemini is now configurable as a server configuration property.
+* The default content type for resources, used when one cannot be detected, is
+now configurable as a server configuration property.
 
 ## v0.3.0